fix ldm half precision

+++ ldm/modules/distributions/distributions.py
@@ -33,7 +33,7 @@ class DiagonalGaussianDistribution(object):
             self.var = self.std = torch.zeros_like(self.mean).to(device=self.parameters.device)
 
     def sample(self):
-        x = self.mean + self.std * torch.randn(self.mean.shape).to(device=self.parameters.device)
+        x = self.mean + self.std * torch.randn(self.mean.shape).to(device=self.parameters.device, dtype=self.parameters.dtype)
         return x
 
     def kl(self, other=None):
